# Entrega/Docker/swagger/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: GeoSolar
  description: GeoSolar
  version: 1.0.0
servers:
- url: http://localhost:5001/api/v1/
  description: Local server
paths:
...
  /users/by-email/{email}:
    get:
      tags:
      - Users
      summary: Get User by Email
      operationId: get_user_by_email
      parameters:
      - name: email
        in: path
        description: User email
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: email
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseUser"
        "400":
          description: Invalid email supplied
        "404":
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
      x-openapi-router-controller: swagger_server.controllers.users_controller
    put:
      tags:
      - Users
      summary: Update User by Email
      operationId: update_user_by_email
      parameters:
      - name: email
        in: path
        description: User email
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: email
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createUser"
        required: true
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseUser"
        "400":
          description: Invalid input
        "404":
          description: User not found
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
      x-openapi-router-controller: swagger_server.controllers.users_controller
    delete:
      tags:
      - Users
      summary: Delete User by Email
      operationId: delete_user_by_email
      parameters:
      - name: email
        in: path
        description: User email
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: email
      responses:
        "204":
          description: User deleted successfully (No Content)
        "400":
          description: Invalid ID supplied
        "404":
          description: User not found
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
      x-openapi-router-controller: swagger_server.controllers.users_controller
    patch:
      tags:
      - Users
      summary: Partially update User by Email
      operationId: patch_user_by_email
      parameters:
      - name: email
        in: path
        description: User email
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: email
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/patchUser"
        required: true
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseUser"
        "400":
          description: Invalid input
        "404":
          description: User not found
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
      x-openapi-router-controller: swagger_server.controllers.users_controller
...
components:
  schemas:
    baseUser:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    responseUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        is_active:
          type: boolean
          example: true
        last_login:
          type: string
          format: date-time
          example: 2025-04-26T00:00:00Z
        created_at:
          type: string
          format: date-time
          example: 2025-04-26T00:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2025-04-26T00:00:00Z
      example:
        is_active: true
        updated_at: 2025-04-26T00:00:00Z
        last_login: 2025-04-26T00:00:00Z
        created_at: 2025-04-26T00:00:00Z
        id: 1
      allOf:
      - $ref: "#/components/schemas/baseUser"
    createUser:
      required:
      - email
      - password
      - username
      allOf:
      - $ref: "#/components/schemas/baseUser"
    patchUser:
      type: object
      properties:
        is_active:
          type: boolean
          example: true
        last_login:
          type: string
          format: date-time
          example: 2025-04-26T00:00:00Z
      allOf:
      - $ref: "#/components/schemas/baseUser"
...
    error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...